refactor(phase5): decompose health.ts into concern-cluster modules
- health: split health.ts (2654 LOC) into health/{improve-metrics,task-runs,windows,llm-usage,metrics,advisories}.ts
- akmHealth remains a thin orchestrator; importers point at the real modules (no re-export barrel)

18 files changed, 2588 insertions(+), 2348 deletions(-) (net +240 from file-boundary tax + added pure-fn tests)
