+++ tests/health-checks-characterization.test.ts
@@ -2,7 +2,8 @@ import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { akmHealth, type HealthCheckResult } from "../src/commands/health";
+import { akmHealth } from "../src/commands/health";
+import type { HealthCheckResult } from "../src/commands/health/types";
 import { appendEvent } from "../src/core/events";
 import { openStateDatabase } from "../src/core/state-db";
 import type { SessionLogEntry } from "../src/integrations/session-logs";
