+++ tests/health-html-report.test.ts
@@ -5,8 +5,9 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import fs from "node:fs";
 import path from "node:path";
-import { type AkmHealthResult, akmHealth } from "../src/commands/health";
+import { akmHealth } from "../src/commands/health";
 import { buildHealthHtmlReplacements, type HealthHtmlReportOptions } from "../src/commands/health/html-report";
+import type { AkmHealthResult } from "../src/commands/health/types";
 import type { AkmImproveResult } from "../src/commands/improve/improve";
 import { appendEvent } from "../src/core/events";
 import { openStateDatabase } from "../src/core/state-db";
